Separate investigation presentation rules and reuse shared helpers
